+++ CHANGELOG.md
@@ -6,6 +6,25 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ### Unreleased
 
+### [3.0.0-alpha.12] - 2026-04-13
+
+#### Added
+
+- toml backend (#49)
+- docker compose support + env var config overrides (#40)
+
+#### Changed
+
+- remove local codeql config (#43)
+- update permissions to be explicit (#42)
+
+#### Other
+
+- add: TOML stores for group, nameserver, permission, session (#47)
+- routes file organization (#46)
+- group factories (#44)
+
+
 - move mysql teardown/disconnect into mysql classes
 - fix: don't log sensitive information
 - add: TOML stores for group, nameserver, permission, session
@@ -86,3 +105,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 [3.0.0-alpha.9]: https://github.com/NicTool/api/releases/tag/v3.0.0-alpha.9
 [3.0.0-alpha.10]: https://github.com/NicTool/api/releases/tag/v3.0.0-alpha.10
 [3.0.0-alpha.11]: https://github.com/NicTool/api/releases/tag/v3.0.0-alpha.11
+[3.0.0-alpha.12]: https://github.com/NicTool/api/releases/tag/v3.0.0-alpha.12